Run TypeScript when task run
Uses tsc provided by the web tools addin if the files are provided.
Otherwise tsc will be run and it is assumed tsc is on the path.
The web tools addin includes Node and tsc.js and these will be
used if they are found. The web tools addin started including these
files in Visual Studio for Mac 7.5.